Add connect() method for convenience
Implement a connect() method. When this method is called, CyTI simply
connects to the first calculator it finds and returns the resulting
Calculator. This is especially useful when using CyTI in the REPL and
you only have one calculator connected to your computer. Actual scripts,
of course, should generally prefer using find_connections().