Remove get_bytes and send_bytes
I've never really been able to get these functions to work, and I'm
honestly not entirely sure that I can. I think that this sort of
communication may need to take place at a lower level to work properly.
Plus, it doesn't seem particularly useful, given that the end goal of
CyTI is to never require such functionality in the first place. So, with
all that in mind, remove these methods in favor of using higher-level
alternatives.