Python int to TI8x real conversion
Create functions to easily convert numbers in Python to TI-8x real
variables that can be dealt with just like variables received from the
calculator. This function does require a name for the variable as well
as a CyTI calculator object so that the exact type can be encoded into
the variable.