Improved strings for Variable[Request]
Change the string representations of these types to include the name of
the variable as well as the calculator and type name associated with
each. As part of this, there are now two-way mappings between type codes
and strings. For example, accessing the dict with [0] will return "real"
and accessing with ["real"] will return 0. This should also make dynamic
variable creation easier.