Variable creation
Allow creation of TI variables. The additional fields in VarEntry were
added to the wrapper type and are set to default values as part of the
creation process. The data for the variable is contained in a Cython
array that backs the publicly accessible data array as well as the
underlying VarEntry, making it easy to change what will be sent to the
calculator from Python code.