Pseudotype support for numbers and lists
Implement pseudotypes for real/complex numbers and lists. This allows a
user to retrieve/send/delete a number or list without needing to know
whether it is real or complex on the calculator. This is somewhat basic,
and each type is simply tried until one works (with the exception of
send). This can probably be improved in delete.

As part of this change, the old "list" variable type is now called
"real_list". The "list" type is now a pseudotype for both real and
complex list types.

Also, similar to get, send now allows the use of an integer type code
instead of a string.