Make raw type conversion methods "private"
Add underscores to the beginnings of the names of these methods to
indicate that they are not necessarily intended for public use.
Conversions should use the to_python and to_cyti methods instead. If
one of those methods isn't having the intended effect, that should be
considered a bug and fixed.