Use getattr to identify complex numbers
It turns out that all numbers in Python appear to have the "imag"
attribute (after all, they are technically complex). So, to distinguish
complex numbers from real numbers, it is necessary to use *getattr*
instead of hasattr and check whether the imaginary part of a number is
0. This fix implements that so that automatic Python -> CyTI type
conversion works better.